sui_rpc_api::grpc::alpha::proof_service_proto::event_service_server

Trait EventService

pub trait EventService:
    Send
    + Sync
    + 'static {
    // Required method
    fn list_authenticated_events<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ListAuthenticatedEventsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ListAuthenticatedEventsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with EventServiceServer.

Required Methods

fn list_authenticated_events<'life0, 'async_trait>( &'life0 self, request: Request<ListAuthenticatedEventsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ListAuthenticatedEventsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns authenticated events for a given event stream ID.