sui_rpc_api::proto::node
Module v2

Modules

effects_finality

Nested message and enum types in EffectsFinality.

node_service_client

Generated client implementations.

node_service_server

Generated server implementations.

Structs

BalanceChange

The delta, or change, in balance for an address for a particular Coin type.

EffectsFinality

Indicates the finality of the executed transaction.

ExecuteTransactionRequest

Request message for NodeService.ExecuteTransaction.

ExecuteTransactionResponse

Response message for NodeService.ExecuteTransaction.

FullCheckpointObject

An object used by or produced from a transaction.

FullCheckpointTransaction

A transaction, with all of its inputs and outputs.

GetCheckpointRequest

Request message for NodeService.GetCheckpoint.

GetCheckpointResponse

Response message for NodeService.GetCheckpoint.

GetCommitteeRequest

Request message for NodeService.GetCommittee.

GetCommitteeResponse

Response message for NodeService.GetCommittee.

GetFullCheckpointRequest

Request message for NodeService.GetFullCheckpoint.

GetFullCheckpointResponse

Response message for NodeService.GetFullCheckpoint.

GetNodeInfoRequest

Request message for NodeService.GetNodeInfo.

GetNodeInfoResponse

Response message for NodeService.GetNodeInfo.

GetObjectRequest

Request message for NodeService.GetObject.

GetObjectResponse

Response message for NodeService.GetObject.

GetTransactionRequest

Request message for NodeService.GetTransaction.

GetTransactionResponse

Response message for NodeService.GetTransaction.

Constants

FILE_DESCRIPTOR_SET

Byte encoded FILE_DESCRIPTOR_SET.