sui_rpc_api::proto::node::v2::node_service_server
Trait NodeService

pub trait NodeService:
    Send
    + Sync
    + 'static {
    // Required methods
    fn get_node_info<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetNodeInfoRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetNodeInfoResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_committee<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetCommitteeRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetCommitteeResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_object<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetObjectRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetObjectResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_transaction<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetTransactionRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetTransactionResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_checkpoint<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetCheckpointRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetCheckpointResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_full_checkpoint<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetFullCheckpointRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetFullCheckpointResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn execute_transaction<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ExecuteTransactionRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ExecuteTransactionResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with NodeServiceServer.

Required Methods

fn get_node_info<'life0, 'async_trait>( &'life0 self, request: Request<GetNodeInfoRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetNodeInfoResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Query a node for information about its current state.

fn get_committee<'life0, 'async_trait>( &'life0 self, request: Request<GetCommitteeRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetCommitteeResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Request the validator committee for a particular epoch or for the current epoch.

fn get_object<'life0, 'async_trait>( &'life0 self, request: Request<GetObjectRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetObjectResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Request information for the specified object.

Use this API to request an object by its ObjectId. The version of the object returned is dependent on if you request a specific version. If you do not request a specific version (GetObjectRequest.version is None), then the most recent version (if the object is live) is returned. If you do request a version, that version is returned if it historically existed, is available, and has not been pruned.

Due to storage limitations, implementers of this service might prune older historical data, which can limit the data availability of this API. To determine the data availability range for historical objects, clients can look at GetNodeInfoResponse.lowest_available_checkpoint_objects to see the lowest checkpoint for which historical object data is available.

fn get_transaction<'life0, 'async_trait>( &'life0 self, request: Request<GetTransactionRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetTransactionResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Request information for the specified transaction.

Use this API to request information about a transaction by its digest.

Due to storage limitations, implementers of this service might prune older historical data, which can limit the data availability of this API. To determine the data availability range for historical transactions, clients can look at GetNodeInfoResponse.lowest_available_checkpoint to see the lowest checkpoint for which historical transaction data is available.

fn get_checkpoint<'life0, 'async_trait>( &'life0 self, request: Request<GetCheckpointRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetCheckpointResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Request information for the specified checkpoint.

Use this API to request information about a checkpoint either by its digest or its sequence number (height).

Due to storage limitations, implementers of this service might prune older historical data, which can limit the data availability of this API. To determine the data availability range for historical checkpoints, clients can look at GetNodeInfoResponse.lowest_available_checkpoint to see the lowest checkpoint for which historical checkpoint data is available.

fn get_full_checkpoint<'life0, 'async_trait>( &'life0 self, request: Request<GetFullCheckpointRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetFullCheckpointResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Request information for the entirety of the specified checkpoint.

Use this API to request information about a checkpoint either by its digest or its sequence number (height). In particular, you can use this API to request information about all the transactions included in a checkpoint, as well as their input and output objects.

Due to storage limitations, implementers of this service might prune older historical data, which can limit the data availability of this API. To determine the data availability range for historical checkpoints, clients can look at GetNodeInfoResponse.lowest_available_checkpoint to see the lowest checkpoint for which historical checkpoint/transaction data is available and GetNodeInfoResponse.lowest_available_checkpoint_objects for which historical object data is available.

fn execute_transaction<'life0, 'async_trait>( &'life0 self, request: Request<ExecuteTransactionRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ExecuteTransactionResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Request that the provided transaction be relayed to the validator set for execution and inclusion in the blockchain.