sui_rpc_api::proto::rpc
Module v2alpha

Modules

datatype_descriptor

Nested message and enum types in DatatypeDescriptor.

dynamic_field

Nested message and enum types in DynamicField.

function_descriptor

Nested message and enum types in FunctionDescriptor.

live_data_service_client

Generated client implementations.

live_data_service_server

Generated server implementations.

move_package_service_client

Generated client implementations.

move_package_service_server

Generated server implementations.

open_signature

Nested message and enum types in OpenSignature.

open_signature_body

Nested message and enum types in OpenSignatureBody.

signature_verification_service_client

Generated client implementations.

signature_verification_service_server

Generated server implementations.

simulate_transaction_request

Nested message and enum types in SimulateTransactionRequest.

subscription_service_client

Generated client implementations.

subscription_service_server

Generated server implementations.

Structs

Balance

Balance information for a specific coin type.

CoinMetadata

Metadata for a coin type

CoinTreasury

Information about a coin type’s 0x2::coin::TreasuryCap and its total available supply

CommandOutput

CommandResult

An intermediate result/output from the execution of a single command

DatatypeDescriptor

Describes a Move Datatype.

DynamicField

FieldDescriptor

Descriptor of a field that belongs to a struct or enum variant

FunctionDescriptor

Descriptor of a Move function

GetBalanceRequest

Request message for LiveDataService.GetBalance.

GetBalanceResponse

Response message for LiveDataService.GetBalance. Return the total coin balance for one coin type, owned by the address owner.

GetCoinInfoRequest

Request message for NodeService.GetCoinInfo.

GetCoinInfoResponse

Response message for NodeService.GetCoinInfo.

GetDatatypeRequest

GetDatatypeResponse

GetFunctionRequest

GetFunctionResponse

GetModuleRequest

GetModuleResponse

GetPackageRequest

GetPackageResponse

ListBalancesRequest

Request message for LiveDataService.ListBalances.

ListBalancesResponse

Response message for LiveDataService.ListBalances. Return the total coin balance for all coin types, owned by the address owner.

ListDynamicFieldsRequest

Request message for NodeService.ListDynamicFields

ListDynamicFieldsResponse

Response message for NodeService.ListDynamicFields

ListOwnedObjectsRequest

ListOwnedObjectsResponse

Module

A Move Module.

OpenSignature

Representation of a type signature that could appear as a function parameter or return value.

OpenSignatureBody

Representation of a type signature that could appear as a field type for a struct or enum

OwnedObject

Package

A Move Package

RegulatedCoinMetadata

Information about a regulated coin, which indicates that it makes use of the transfer deny list.

SimulateTransactionRequest

SimulateTransactionResponse

SubscribeCheckpointsRequest

Request message for SubscriptionService.SubscribeCheckpoints

SubscribeCheckpointsResponse

Response message for SubscriptionService.SubscribeCheckpoints

TypeParameter

A generic type parameter used in the declaration of a struct or enum.

VariantDescriptor

Descriptor of an enum variant

VerifySignatureRequest

VerifySignatureResponse

Enums

Ability

An Ability classifies what operations are permitted for a given type

Constants

FILE_DESCRIPTOR_SET

Byte encoded FILE_DESCRIPTOR_SET.