sui_rpc_api::proto::rpc
Module v2beta

Modules

argument

Nested message and enum types in Argument.

changed_object

Nested message and enum types in ChangedObject.

checkpoint_commitment

Nested message and enum types in CheckpointCommitment.

command

Nested message and enum types in Command.

command_argument_error

Nested message and enum types in CommandArgumentError.

end_of_epoch_transaction_kind

Nested message and enum types in EndOfEpochTransactionKind.

execution_error

Nested message and enum types in ExecutionError.

execution_time_observation

Nested message and enum types in ExecutionTimeObservation.

get_checkpoint_request

Nested message and enum types in GetCheckpointRequest.

input

Nested message and enum types in Input.

ledger_service_client

Generated client implementations.

ledger_service_server

Generated server implementations.

owner

Nested message and enum types in Owner.

package_upgrade_error

Nested message and enum types in PackageUpgradeError.

transaction_execution_service_client

Generated client implementations.

transaction_execution_service_server

Generated server implementations.

transaction_expiration

Nested message and enum types in TransactionExpiration.

transaction_finality

Nested message and enum types in TransactionFinality.

transaction_kind

Nested message and enum types in TransactionKind.

type_argument_error

Nested message and enum types in TypeArgumentError.

unchanged_shared_object

Nested message and enum types in UnchangedSharedObject.

Structs

ActiveJwk

A new JWK.

Argument

An argument to a programmable transaction command.

AuthenticatorStateExpire

Expire old JWKs.

AuthenticatorStateUpdate

Update the set of valid JWKs.

BalanceChange

The delta, or change, in balance for an address for a particular Coin type.

BatchGetObjectsRequest

BatchGetObjectsResponse

BatchGetTransactionsRequest

BatchGetTransactionsResponse

Bcs

Bcs contains an arbitrary type that is serialized using the BCS format as well as a name that identifies the type of the serialized value.

CanceledTransaction

A transaction that was canceled.

ChangeEpoch

System transaction used to change the epoch.

ChangedObject

Input/output state of an object that was changed during execution.

Checkpoint

CheckpointCommitment

A commitment made by a checkpoint.

CheckpointContents

The committed to contents of a checkpoint.

CheckpointSummary

A header for a checkpoint on the Sui blockchain.

CheckpointedTransactionInfo

Transaction information committed to in a checkpoint.

CircomG1

A G1 point.

CircomG2

A G2 point.

Command

A single command in a programmable transaction.

CommandArgumentError

An error with an argument to a command.

ConsensusCommitPrologue

Consensus commit prologue system transaction.

ConsensusDeterminedVersionAssignments

Version assignments performed by consensus.

EndOfEpochData

Data, which when included in a CheckpointSummary, signals the end of an Epoch.

EndOfEpochTransaction

Set of operations run at the end of the epoch to close out the current epoch and start the next one.

EndOfEpochTransactionKind

Operation run at the end of an epoch.

Epoch

Event

An event.

ExecuteTransactionRequest

ExecuteTransactionResponse

Response message for NodeService.ExecuteTransaction.

ExecutedTransaction

ExecutionError

An error that can occur during the execution of a transaction.

ExecutionStatus

The status of an executed transaction.

ExecutionTimeObservation

ExecutionTimeObservations

GasCostSummary

Summary of gas charges.

GasPayment

Payment information for executing a transaction.

GenesisTransaction

The genesis transaction.

GetCheckpointRequest

GetEpochRequest

GetObjectRequest

GetServiceInfoRequest

GetServiceInfoResponse

GetTransactionRequest

Input

An input to a user transaction.

Jwk

A JSON web key.

JwkId

Key to uniquely identify a JWK.

MakeMoveVector

Command to build a Move vector out of a set of individual elements.

MergeCoins

Command to merge multiple coins of the same type into a single coin.

MoveCall

Command to call a Move function.

MoveLocation

Location in Move bytecode where an error occurred.

MoveModule

Module defined by a package.

MoveTable

A message that represents a Move 0x2::table::Table or 0x2::bag::Bag

MultisigAggregatedSignature

Aggregated signature from members of a multisig committee.

MultisigCommittee

A multisig committee.

MultisigMember

A member in a multisig committee.

MultisigMemberPublicKey

Set of valid public keys for multisig committee members.

MultisigMemberSignature

A signature from a member of a multisig committee.

Object

An object on the Sui blockchain.

ObjectReference

Reference to an object.

Owner

Enum of different types of ownership for an object.

PackageUpgradeError

An error with upgrading a package.

PasskeyAuthenticator

A passkey authenticator.

ProgrammableTransaction

A user transaction.

ProtocolConfig

Publish

Command to publish a new Move package.

RandomnessStateUpdate

Randomness update.

SizeError

A size error.

SplitCoins

Command to split a single coin object into multiple coins.

StakeSubsidy

StakingPool

A staking pool embedded in each validator struct in the system state object.

StorageFund

Struct representing the onchain storage fund.

SystemPackage

System package.

SystemParameters

SystemState

Transaction

A transaction.

TransactionEffects

The effects of executing a transaction.

TransactionEvents

Events emitted during the successful execution of a transaction.

TransactionExpiration

A TTL for a transaction.

TransactionFinality

Indicates the finality of the executed transaction.

TransactionKind

Transaction type.

TransferObjects

Command to transfer ownership of a set of objects to an address.

TypeArgumentError

Type argument error.

TypeOrigin

Identifies a struct and the module it was defined in.

UnchangedSharedObject

A shared object that wasn’t changed during execution.

Upgrade

Command to upgrade an already published package.

UpgradeInfo

/ Upgraded package info for the linkage table.

UserSignature

A signature from a user.

Validator

Definition of a Validator in the system contracts

ValidatorAggregatedSignature

/ An aggregated signature from multiple validators.

ValidatorCommittee

The validator set for a particular epoch.

ValidatorCommitteeMember

A member of a validator committee.

ValidatorExecutionTimeObservation

ValidatorReportRecord

ValidatorSet

VersionAssignment

Object version assignment from consensus.

ZkLoginAuthenticator

A zklogin authenticator.

ZkLoginClaim

A claim of the iss in a zklogin proof.

ZkLoginInputs

A zklogin groth16 proof and the required inputs to perform proof verification.

ZkLoginProof

A zklogin groth16 proof.

ZkLoginPublicIdentifier

Public key equivalent for zklogin authenticators.

Enums

SignatureScheme

Flag use to disambiguate the signature schemes supported by Sui.

Constants

FILE_DESCRIPTOR_SET

Byte encoded FILE_DESCRIPTOR_SET.