sui_rpc_api::proto::rpc::v2beta::ledger_service_server
Trait LedgerService

pub trait LedgerService:
    Send
    + Sync
    + 'static {
    // Required methods
    fn get_service_info<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetServiceInfoRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetServiceInfoResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_object<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetObjectRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Object>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn batch_get_objects<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BatchGetObjectsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<BatchGetObjectsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_transaction<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetTransactionRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ExecutedTransaction>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn batch_get_transactions<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BatchGetTransactionsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<BatchGetTransactionsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_checkpoint<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetCheckpointRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Checkpoint>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_epoch<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetEpochRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Epoch>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with LedgerServiceServer.

Required Methods

fn get_service_info<'life0, 'async_trait>( &'life0 self, request: Request<GetServiceInfoRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetServiceInfoResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Query the service for general information about its current state.

fn get_object<'life0, 'async_trait>( &'life0 self, request: Request<GetObjectRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Object>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn batch_get_objects<'life0, 'async_trait>( &'life0 self, request: Request<BatchGetObjectsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<BatchGetObjectsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_transaction<'life0, 'async_trait>( &'life0 self, request: Request<GetTransactionRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ExecutedTransaction>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn batch_get_transactions<'life0, 'async_trait>( &'life0 self, request: Request<BatchGetTransactionsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<BatchGetTransactionsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_checkpoint<'life0, 'async_trait>( &'life0 self, request: Request<GetCheckpointRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Checkpoint>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_epoch<'life0, 'async_trait>( &'life0 self, request: Request<GetEpochRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Epoch>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,