sui_rpc_api::proto::sui::rpc

Module v2

Modules

accumulator_write

Nested message and enum types in AccumulatorWrite.

argument

Nested message and enum types in Argument.

changed_object

Nested message and enum types in ChangedObject.

checkpoint_commitment

Nested message and enum types in CheckpointCommitment.

clever_error

Nested message and enum types in CleverError.

coin_metadata

Nested message and enum types in CoinMetadata.

coin_treasury

Nested message and enum types in CoinTreasury.

command

Nested message and enum types in Command.

command_argument_error

Nested message and enum types in CommandArgumentError.

datatype_descriptor

Nested message and enum types in DatatypeDescriptor.

dynamic_field

Nested message and enum types in DynamicField.

end_of_epoch_transaction_kind

Nested message and enum types in EndOfEpochTransactionKind.

execution_error

Nested message and enum types in ExecutionError.

execution_time_observation

Nested message and enum types in ExecutionTimeObservation.

function_descriptor

Nested message and enum types in FunctionDescriptor.

funds_withdrawal

Nested message and enum types in FundsWithdrawal.

get_checkpoint_request

Nested message and enum types in GetCheckpointRequest.

get_object_result

Nested message and enum types in GetObjectResult.

get_transaction_result

Nested message and enum types in GetTransactionResult.

input

Nested message and enum types in Input.

ledger_service_client

Generated client implementations.

ledger_service_server

Generated server implementations.

move_package_service_client

Generated client implementations.

move_package_service_server

Generated server implementations.

name_service_client

Generated client implementations.

name_service_server

Generated server implementations.

open_signature

Nested message and enum types in OpenSignature.

open_signature_body

Nested message and enum types in OpenSignatureBody.

owner

Nested message and enum types in Owner.

package_upgrade_error

Nested message and enum types in PackageUpgradeError.

regulated_coin_metadata

Nested message and enum types in RegulatedCoinMetadata.

signature_verification_service_client

Generated client implementations.

signature_verification_service_server

Generated server implementations.

simulate_transaction_request

Nested message and enum types in SimulateTransactionRequest.

state_service_client

Generated client implementations.

state_service_server

Generated server implementations.

subscription_service_client

Generated client implementations.

subscription_service_server

Generated server implementations.

transaction_execution_service_client

Generated client implementations.

transaction_execution_service_server

Generated server implementations.

transaction_expiration

Nested message and enum types in TransactionExpiration.

transaction_kind

Nested message and enum types in TransactionKind.

type_argument_error

Nested message and enum types in TypeArgumentError.

unchanged_consensus_object

Nested message and enum types in UnchangedConsensusObject.

user_signature

Nested message and enum types in UserSignature.

Structs

AccumulatorWrite

ActiveJwk

A new JWK.

AddressAliasesVersion

Argument

An argument to a programmable transaction command.

AuthenticatorStateExpire

Expire old JWKs.

AuthenticatorStateUpdate

Update the set of valid JWKs.

Balance

Balance information for a specific coin type.

BalanceChange

The delta, or change, in balance for an address for a particular Coin type.

BatchGetObjectsRequest

BatchGetObjectsResponse

BatchGetTransactionsRequest

BatchGetTransactionsResponse

Bcs

Bcs contains an arbitrary type that is serialized using the BCS format as well as a name that identifies the type of the serialized value.

CanceledTransaction

A transaction that was canceled.

ChangeEpoch

System transaction used to change the epoch.

ChangedObject

Input/output state of an object that was changed during execution.

Checkpoint

CheckpointCommitment

A commitment made by a checkpoint.

CheckpointContents

The committed to contents of a checkpoint.

CheckpointSummary

A header for a checkpoint on the Sui blockchain.

CheckpointedTransactionInfo

Transaction information committed to in a checkpoint.

CircomG1

A G1 point.

CircomG2

A G2 point.

CleverError

CoinDenyListError

CoinMetadata

Metadata for a coin type

CoinTreasury

Information about a coin type’s 0x2::coin::TreasuryCap and its total available supply

Command

A single command in a programmable transaction.

CommandArgumentError

An error with an argument to a command.

CommandOutput

CommandResult

An intermediate result/output from the execution of a single command

CongestedObjects

Set of objects that were congested, leading to the transaction’s cancellation.

ConsensusCommitPrologue

Consensus commit prologue system transaction.

ConsensusDeterminedVersionAssignments

Version assignments performed by consensus.

DatatypeDescriptor

Describes a Move Datatype.

DynamicField

EndOfEpochData

Data, which when included in a CheckpointSummary, signals the end of an Epoch.

EndOfEpochTransaction

Set of operations run at the end of the epoch to close out the current epoch and start the next one.

EndOfEpochTransactionKind

Operation run at the end of an epoch.

Epoch

Event

An event.

ExecuteTransactionRequest

ExecuteTransactionResponse

Response message for NodeService.ExecuteTransaction.

ExecutedTransaction

ExecutionError

An error that can occur during the execution of a transaction.

ExecutionStatus

The status of an executed transaction.

ExecutionTimeObservation

ExecutionTimeObservations

FieldDescriptor

Descriptor of a field that belongs to a struct or enum variant

FunctionDescriptor

Descriptor of a Move function

FundsWithdrawal

GasCostSummary

Summary of gas charges.

GasPayment

Payment information for executing a transaction.

GenesisTransaction

The genesis transaction.

GetBalanceRequest

Request message for LiveDataService.GetBalance.

GetBalanceResponse

Response message for LiveDataService.GetBalance. Return the total coin balance for one coin type, owned by the address owner.

GetCheckpointRequest

GetCheckpointResponse

GetCoinInfoRequest

Request message for NodeService.GetCoinInfo.

GetCoinInfoResponse

Response message for NodeService.GetCoinInfo.

GetDatatypeRequest

GetDatatypeResponse

GetEpochRequest

GetEpochResponse

GetFunctionRequest

GetFunctionResponse

GetObjectRequest

GetObjectResponse

GetObjectResult

GetPackageRequest

GetPackageResponse

GetServiceInfoRequest

GetServiceInfoResponse

GetTransactionRequest

GetTransactionResponse

GetTransactionResult

IndexError

Input

An input to a user transaction.

Jwk

A JSON web key.

JwkId

Key to uniquely identify a JWK.

Linkage

Upgraded package info for the linkage table.

ListBalancesRequest

Request message for LiveDataService.ListBalances.

ListBalancesResponse

Response message for LiveDataService.ListBalances. Return the total coin balance for all coin types, owned by the address owner.

ListDynamicFieldsRequest

Request message for NodeService.ListDynamicFields

ListDynamicFieldsResponse

Response message for NodeService.ListDynamicFields

ListOwnedObjectsRequest

ListOwnedObjectsResponse

ListPackageVersionsRequest

ListPackageVersionsResponse

LookupNameRequest

LookupNameResponse

MakeMoveVector

Command to build a Move vector out of a set of individual elements.

MergeCoins

Command to merge multiple coins of the same type into a single coin.

Module

A Move Module.

MoveAbort

MoveCall

Command to call a Move function.

MoveLocation

Location in Move bytecode where an error occurred.

MoveTable

A message that represents a Move 0x2::table::Table or 0x2::bag::Bag

MultisigAggregatedSignature

Aggregated signature from members of a multisig committee.

MultisigCommittee

A multisig committee.

MultisigMember

A member in a multisig committee.

MultisigMemberPublicKey

Set of valid public keys for multisig committee members.

MultisigMemberSignature

A signature from a member of a multisig committee.

NameRecord

Object

An object on the Sui blockchain.

ObjectReference

Reference to an object.

ObjectSet

Set of Objects

OpenSignature

Representation of a type signature that could appear as a function parameter or return value.

OpenSignatureBody

Representation of a type signature that could appear as a field type for a struct or enum

Owner

Enum of different types of ownership for an object.

Package

A Move Package

PackageUpgradeError

An error with upgrading a package.

PackageVersion

A simplified representation of a package version

PasskeyAuthenticator

A passkey authenticator.

ProgrammableTransaction

A user transaction.

ProtocolConfig

Publish

Command to publish a new Move package.

RandomnessStateUpdate

Randomness update.

RegulatedCoinMetadata

Information about a regulated coin, which indicates that it makes use of the transfer deny list.

ReverseLookupNameRequest

ReverseLookupNameResponse

SimpleSignature

Either an ed25519, secp256k1 or secp256r1 signature

SimulateTransactionRequest

SimulateTransactionResponse

SizeError

A size error.

SplitCoins

Command to split a single coin object into multiple coins.

StakeSubsidy

StakingPool

A staking pool embedded in each validator struct in the system state object.

StorageFund

Struct representing the onchain storage fund.

SubscribeCheckpointsRequest

Request message for SubscriptionService.SubscribeCheckpoints

SubscribeCheckpointsResponse

Response message for SubscriptionService.SubscribeCheckpoints

SystemPackage

System package.

SystemParameters

SystemState

Transaction

A transaction.

TransactionEffects

The effects of executing a transaction.

TransactionEvents

Events emitted during the successful execution of a transaction.

TransactionExpiration

A TTL for a transaction.

TransactionKind

Transaction type.

TransferObjects

Command to transfer ownership of a set of objects to an address.

TypeArgumentError

Type argument error.

TypeOrigin

Identifies a struct and the module it was defined in.

TypeParameter

A generic type parameter used in the declaration of a struct or enum.

UnchangedConsensusObject

A consensus object that wasn’t changed during execution.

Upgrade

Command to upgrade an already published package.

UserSignature

A signature from a user.

Validator

Definition of a Validator in the system contracts

ValidatorAggregatedSignature

/ An aggregated signature from multiple validators.

ValidatorCommittee

The validator set for a particular epoch.

ValidatorCommitteeMember

A member of a validator committee.

ValidatorExecutionTimeObservation

ValidatorReportRecord

ValidatorSet

VariantDescriptor

Descriptor of an enum variant

VerifySignatureRequest

VerifySignatureResponse

VersionAssignment

Object version assignment from consensus.

ZkLoginAuthenticator

A zklogin authenticator.

ZkLoginClaim

A claim of the iss in a zklogin proof.

ZkLoginInputs

A zklogin groth16 proof and the required inputs to perform proof verification.

ZkLoginProof

A zklogin groth16 proof.

ZkLoginPublicIdentifier

Public key equivalent for zklogin authenticators.

Enums

Ability

An Ability classifies what operations are permitted for a given type

ErrorReason

SignatureScheme

Flag use to disambiguate the signature schemes supported by Sui.

Constants

FILE_DESCRIPTOR_SET

Byte encoded FILE_DESCRIPTOR_SET.