sui_rpc_api::proto::sui::rpc::v2::subscription_service_server

Trait SubscriptionService

pub trait SubscriptionService:
    Send
    + Sync
    + 'static {
    // Provided methods
    fn subscribe_checkpoints<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SubscribeCheckpointsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<SubscribeCheckpointsResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn subscribe_transactions<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SubscribeTransactionsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<SubscribeTransactionsResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn subscribe_events<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SubscribeEventsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<SubscribeEventsResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
}

Generated trait containing gRPC methods that should be implemented for use with SubscriptionServiceServer.

Provided Methods

fn subscribe_checkpoints<'life0, 'async_trait>( &'life0 self, request: Request<SubscribeCheckpointsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<SubscribeCheckpointsResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Subscribe to the stream of checkpoints.

The stream begins at the latest executed checkpoint as seen by the server and yields checkpoints matching the filter as they are executed. A checkpoint matches if any transaction it contains satisfies the filter.

fn subscribe_transactions<'life0, 'async_trait>( &'life0 self, request: Request<SubscribeTransactionsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<SubscribeTransactionsResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Subscribe to the stream of transactions.

The stream begins at the latest executed checkpoint as seen by the server and yields transactions matching the filter as they are executed.

fn subscribe_events<'life0, 'async_trait>( &'life0 self, request: Request<SubscribeEventsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<SubscribeEventsResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Subscribe to the stream of events.

The stream begins at the latest executed checkpoint as seen by the server and yields events matching the filter as they are emitted.