sui_rpc_api::proto::sui::rpc::v2alpha::ledger_service_server

Trait LedgerService

pub trait LedgerService:
    Send
    + Sync
    + 'static {
    // Provided methods
    fn list_checkpoints<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ListCheckpointsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<ListCheckpointsResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn list_transactions<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ListTransactionsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<ListTransactionsResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn list_events<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ListEventsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<ListEventsResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
}

Generated trait containing gRPC methods that should be implemented for use with LedgerServiceServer.

Provided Methods

fn list_checkpoints<'life0, 'async_trait>( &'life0 self, request: Request<ListCheckpointsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<ListCheckpointsResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

List checkpoints matching the provided filters.

Checkpoints are returned in ascending or descending checkpoint sequence number order according to the query options ordering. A checkpoint matches if any transaction it contains satisfies the filter.

fn list_transactions<'life0, 'async_trait>( &'life0 self, request: Request<ListTransactionsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<ListTransactionsResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

List transactions matching the provided filters.

Transactions are returned in ascending or descending transaction sequence order according to the query options ordering.

fn list_events<'life0, 'async_trait>( &'life0 self, request: Request<ListEventsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<ListEventsResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

List events matching the provided filters.

Events are returned in ascending or descending packed event sequence order according to the query options ordering.