sui_rpc_api::proto::sui::rpc::v2beta2
Struct CheckpointSummary

#[non_exhaustive]
pub struct CheckpointSummary {
    pub bcs: Option<Bcs>,
    pub digest: Option<String>,
    pub epoch: Option<u64>,
    pub sequence_number: Option<u64>,
    pub total_network_transactions: Option<u64>,
    pub content_digest: Option<String>,
    pub previous_digest: Option<String>,
    pub epoch_rolling_gas_cost_summary: Option<GasCostSummary>,
    pub timestamp: Option<Timestamp>,
    pub commitments: Vec<CheckpointCommitment>,
    pub end_of_epoch_data: Option<EndOfEpochData>,
    pub version_specific_data: Option<Bytes>,
}

A header for a checkpoint on the Sui blockchain.

On the Sui network, checkpoints define the history of the blockchain. They are quite similar to the concept of blocks used by other blockchains like Bitcoin or Ethereum. The Sui blockchain, however, forms checkpoints after transaction execution has already happened to provide a certified history of the chain, instead of being formed before execution.

Checkpoints commit to a variety of state, including but not limited to:

The hash of the previous checkpoint.

The set of transaction digests, their corresponding effects digests, as well as the set of user signatures that authorized its execution.

The objects produced by a transaction.

The set of live objects that make up the current state of the chain.

On epoch transitions, the next validator committee.

CheckpointSummarys themselves don’t directly include all of the previous information but they are the top-level type by which all the information is committed to transitively via cryptographic hashes included in the summary. CheckpointSummarys are signed and certified by a quorum of the validator committee in a given epoch to allow verification of the chain’s state.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

bcs: Option<Bcs>

This CheckpointSummary serialized as BCS.

digest: Option<String>

The digest of this CheckpointSummary.

epoch: Option<u64>

Epoch that this checkpoint belongs to.

sequence_number: Option<u64>

The height of this checkpoint.

total_network_transactions: Option<u64>

Total number of transactions committed since genesis, including those in this checkpoint.

content_digest: Option<String>

The hash of the CheckpointContents for this checkpoint.

previous_digest: Option<String>

The hash of the previous CheckpointSummary.

This will be None only for the first, or genesis, checkpoint.

epoch_rolling_gas_cost_summary: Option<GasCostSummary>

The running total gas costs of all transactions included in the current epoch so far until this checkpoint.

timestamp: Option<Timestamp>

Timestamp of the checkpoint - number of milliseconds from the Unix epoch Checkpoint timestamps are monotonic, but not strongly monotonic - subsequent checkpoints can have the same timestamp if they originate from the same underlining consensus commit.

commitments: Vec<CheckpointCommitment>

Commitments to checkpoint-specific state.

end_of_epoch_data: Option<EndOfEpochData>

Extra data only present in the final checkpoint of an epoch.

version_specific_data: Option<Bytes>

CheckpointSummary is not an evolvable structure - it must be readable by any version of the code. Therefore, to allow extensions to be added to CheckpointSummary, opaque data can be added to checkpoints, which can be deserialized based on the current protocol version.

Implementations

impl CheckpointSummary

pub fn digest(&self) -> &str

Returns the value of digest, or the default value if digest is unset.

pub fn epoch(&self) -> u64

Returns the value of epoch, or the default value if epoch is unset.

pub fn sequence_number(&self) -> u64

Returns the value of sequence_number, or the default value if sequence_number is unset.

pub fn total_network_transactions(&self) -> u64

Returns the value of total_network_transactions, or the default value if total_network_transactions is unset.

pub fn content_digest(&self) -> &str

Returns the value of content_digest, or the default value if content_digest is unset.

pub fn previous_digest(&self) -> &str

Returns the value of previous_digest, or the default value if previous_digest is unset.

pub fn version_specific_data(&self) -> &[u8]

Returns the value of version_specific_data, or the default value if version_specific_data is unset.

impl CheckpointSummary

pub const BCS_FIELD: &'static MessageField

pub const DIGEST_FIELD: &'static MessageField

pub const EPOCH_FIELD: &'static MessageField

pub const SEQUENCE_NUMBER_FIELD: &'static MessageField

pub const TOTAL_NETWORK_TRANSACTIONS_FIELD: &'static MessageField

pub const CONTENT_DIGEST_FIELD: &'static MessageField

pub const PREVIOUS_DIGEST_FIELD: &'static MessageField

pub const EPOCH_ROLLING_GAS_COST_SUMMARY_FIELD: &'static MessageField

pub const TIMESTAMP_FIELD: &'static MessageField

pub const COMMITMENTS_FIELD: &'static MessageField

pub const END_OF_EPOCH_DATA_FIELD: &'static MessageField

pub const VERSION_SPECIFIC_DATA_FIELD: &'static MessageField

impl CheckpointSummary

pub fn path_builder() -> CheckpointSummaryFieldPathBuilder

impl CheckpointSummary

pub const fn const_default() -> CheckpointSummary

pub fn bcs(&self) -> &Bcs

Returns the value of bcs, or the default value if bcs is unset.

pub fn bcs_opt_mut(&mut self) -> Option<&mut Bcs>

If bcs is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn bcs_mut(&mut self) -> &mut Bcs

Returns a mutable reference to bcs. If the field is unset, it is first initialized with the default value.

pub fn bcs_opt(&self) -> Option<&Bcs>

If bcs is set, returns Some with the value; otherwise returns None.

pub fn set_bcs<T>(&mut self, field: T)
where T: Into<Bcs>,

Sets bcs with the provided value.

pub fn with_bcs<T>(self, field: T) -> CheckpointSummary
where T: Into<Bcs>,

Sets bcs with the provided value.

pub fn digest_opt_mut(&mut self) -> Option<&mut String>

If digest is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn digest_mut(&mut self) -> &mut String

Returns a mutable reference to digest. If the field is unset, it is first initialized with the default value.

pub fn digest_opt(&self) -> Option<&str>

If digest is set, returns Some with the value; otherwise returns None.

pub fn set_digest<T>(&mut self, field: T)
where T: Into<String>,

Sets digest with the provided value.

pub fn with_digest<T>(self, field: T) -> CheckpointSummary
where T: Into<String>,

Sets digest with the provided value.

pub fn epoch_opt_mut(&mut self) -> Option<&mut u64>

If epoch is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn epoch_mut(&mut self) -> &mut u64

Returns a mutable reference to epoch. If the field is unset, it is first initialized with the default value.

pub fn epoch_opt(&self) -> Option<u64>

If epoch is set, returns Some with the value; otherwise returns None.

pub fn set_epoch<T>(&mut self, field: T)
where T: Into<u64>,

Sets epoch with the provided value.

pub fn with_epoch<T>(self, field: T) -> CheckpointSummary
where T: Into<u64>,

Sets epoch with the provided value.

pub fn sequence_number_opt_mut(&mut self) -> Option<&mut u64>

If sequence_number is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn sequence_number_mut(&mut self) -> &mut u64

Returns a mutable reference to sequence_number. If the field is unset, it is first initialized with the default value.

pub fn sequence_number_opt(&self) -> Option<u64>

If sequence_number is set, returns Some with the value; otherwise returns None.

pub fn set_sequence_number<T>(&mut self, field: T)
where T: Into<u64>,

Sets sequence_number with the provided value.

pub fn with_sequence_number<T>(self, field: T) -> CheckpointSummary
where T: Into<u64>,

Sets sequence_number with the provided value.

pub fn total_network_transactions_opt_mut(&mut self) -> Option<&mut u64>

If total_network_transactions is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn total_network_transactions_mut(&mut self) -> &mut u64

Returns a mutable reference to total_network_transactions. If the field is unset, it is first initialized with the default value.

pub fn total_network_transactions_opt(&self) -> Option<u64>

If total_network_transactions is set, returns Some with the value; otherwise returns None.

pub fn set_total_network_transactions<T>(&mut self, field: T)
where T: Into<u64>,

Sets total_network_transactions with the provided value.

pub fn with_total_network_transactions<T>(self, field: T) -> CheckpointSummary
where T: Into<u64>,

Sets total_network_transactions with the provided value.

pub fn content_digest_opt_mut(&mut self) -> Option<&mut String>

If content_digest is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn content_digest_mut(&mut self) -> &mut String

Returns a mutable reference to content_digest. If the field is unset, it is first initialized with the default value.

pub fn content_digest_opt(&self) -> Option<&str>

If content_digest is set, returns Some with the value; otherwise returns None.

pub fn set_content_digest<T>(&mut self, field: T)
where T: Into<String>,

Sets content_digest with the provided value.

pub fn with_content_digest<T>(self, field: T) -> CheckpointSummary
where T: Into<String>,

Sets content_digest with the provided value.

pub fn previous_digest_opt_mut(&mut self) -> Option<&mut String>

If previous_digest is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn previous_digest_mut(&mut self) -> &mut String

Returns a mutable reference to previous_digest. If the field is unset, it is first initialized with the default value.

pub fn previous_digest_opt(&self) -> Option<&str>

If previous_digest is set, returns Some with the value; otherwise returns None.

pub fn set_previous_digest<T>(&mut self, field: T)
where T: Into<String>,

Sets previous_digest with the provided value.

pub fn with_previous_digest<T>(self, field: T) -> CheckpointSummary
where T: Into<String>,

Sets previous_digest with the provided value.

pub fn epoch_rolling_gas_cost_summary(&self) -> &GasCostSummary

Returns the value of epoch_rolling_gas_cost_summary, or the default value if epoch_rolling_gas_cost_summary is unset.

pub fn epoch_rolling_gas_cost_summary_opt_mut( &mut self, ) -> Option<&mut GasCostSummary>

If epoch_rolling_gas_cost_summary is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn epoch_rolling_gas_cost_summary_mut(&mut self) -> &mut GasCostSummary

Returns a mutable reference to epoch_rolling_gas_cost_summary. If the field is unset, it is first initialized with the default value.

pub fn epoch_rolling_gas_cost_summary_opt(&self) -> Option<&GasCostSummary>

If epoch_rolling_gas_cost_summary is set, returns Some with the value; otherwise returns None.

pub fn set_epoch_rolling_gas_cost_summary<T>(&mut self, field: T)
where T: Into<GasCostSummary>,

Sets epoch_rolling_gas_cost_summary with the provided value.

pub fn with_epoch_rolling_gas_cost_summary<T>( self, field: T, ) -> CheckpointSummary
where T: Into<GasCostSummary>,

Sets epoch_rolling_gas_cost_summary with the provided value.

pub fn timestamp_opt_mut(&mut self) -> Option<&mut Timestamp>

If timestamp is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn timestamp_mut(&mut self) -> &mut Timestamp

Returns a mutable reference to timestamp. If the field is unset, it is first initialized with the default value.

pub fn timestamp_opt(&self) -> Option<&Timestamp>

If timestamp is set, returns Some with the value; otherwise returns None.

pub fn set_timestamp<T>(&mut self, field: T)
where T: Into<Timestamp>,

Sets timestamp with the provided value.

pub fn with_timestamp<T>(self, field: T) -> CheckpointSummary
where T: Into<Timestamp>,

Sets timestamp with the provided value.

pub fn commitments(&self) -> &[CheckpointCommitment]

Returns the value of commitments, or the default value if commitments is unset.

pub fn commitments_mut(&mut self) -> &mut Vec<CheckpointCommitment>

Returns a mutable reference to commitments. If the field is unset, it is first initialized with the default value.

pub fn set_commitments(&mut self, field: Vec<CheckpointCommitment>)

Sets commitments with the provided value.

pub fn with_commitments( self, field: Vec<CheckpointCommitment>, ) -> CheckpointSummary

Sets commitments with the provided value.

pub fn end_of_epoch_data(&self) -> &EndOfEpochData

Returns the value of end_of_epoch_data, or the default value if end_of_epoch_data is unset.

pub fn end_of_epoch_data_opt_mut(&mut self) -> Option<&mut EndOfEpochData>

If end_of_epoch_data is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn end_of_epoch_data_mut(&mut self) -> &mut EndOfEpochData

Returns a mutable reference to end_of_epoch_data. If the field is unset, it is first initialized with the default value.

pub fn end_of_epoch_data_opt(&self) -> Option<&EndOfEpochData>

If end_of_epoch_data is set, returns Some with the value; otherwise returns None.

pub fn set_end_of_epoch_data<T>(&mut self, field: T)
where T: Into<EndOfEpochData>,

Sets end_of_epoch_data with the provided value.

pub fn with_end_of_epoch_data<T>(self, field: T) -> CheckpointSummary
where T: Into<EndOfEpochData>,

Sets end_of_epoch_data with the provided value.

pub fn version_specific_data_opt(&self) -> Option<&[u8]>

If version_specific_data is set, returns Some with the value; otherwise returns None.

pub fn set_version_specific_data<T>(&mut self, field: T)
where T: Into<Bytes>,

Sets version_specific_data with the provided value.

pub fn with_version_specific_data<T>(self, field: T) -> CheckpointSummary
where T: Into<Bytes>,

Sets version_specific_data with the provided value.