sui_rpc_api::proto::sui::rpc::v2beta2
Struct TransactionEffects

#[non_exhaustive]
pub struct TransactionEffects {

    pub bcs: Option<Bcs>,
    pub digest: Option<String>,
    pub version: Option<i32>,
    pub status: Option<ExecutionStatus>,
    pub epoch: Option<u64>,
    pub gas_used: Option<GasCostSummary>,
    pub transaction_digest: Option<String>,
    pub gas_object: Option<ChangedObject>,
    pub events_digest: Option<String>,
    pub dependencies: Vec<String>,
    pub lamport_version: Option<u64>,
    pub changed_objects: Vec<ChangedObject>,
    pub unchanged_consensus_objects: Vec<UnchangedConsensusObject>,
    pub auxiliary_data_digest: Option<String>,

}

The effects of executing a transaction.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

bcs: Option<Bcs>

This TransactionEffects serialized as BCS.

digest: Option<String>

The digest of this TransactionEffects.

version: Option<i32>

Version of this TransactionEffects.

status: Option<ExecutionStatus>

The status of the execution.

epoch: Option<u64>

The epoch when this transaction was executed.

gas_used: Option<GasCostSummary>

The gas used by this transaction.

transaction_digest: Option<String>

The transaction digest.

gas_object: Option<ChangedObject>

Information about the gas object. Also present in the changed_objects vector.

System transaction that don’t require gas will leave this as None.

events_digest: Option<String>

The digest of the events emitted during execution, can be None if the transaction does not emit any event.

dependencies: Vec<String>

The set of transaction digests this transaction depends on.

lamport_version: Option<u64>

The version number of all the written objects (excluding packages) by this transaction.

changed_objects: Vec<ChangedObject>

Objects whose state are changed by this transaction.

unchanged_consensus_objects: Vec<UnchangedConsensusObject>

Consensus objects that are not mutated in this transaction. Unlike owned objects, read-only consensus objects’ version are not committed in the transaction, and in order for a node to catch up and execute it without consensus sequencing, the version needs to be committed in the effects.

auxiliary_data_digest: Option<String>

Auxiliary data that are not protocol-critical, generated as part of the effects but are stored separately. Storing it separately allows us to avoid bloating the effects with data that are not critical. It also provides more flexibility on the format and type of the data.

Implementations

impl TransactionEffects

pub fn digest(&self) -> &str

Returns the value of digest, or the default value if digest is unset.

pub fn version(&self) -> i32

Returns the value of version, or the default value if version is unset.

pub fn epoch(&self) -> u64

Returns the value of epoch, or the default value if epoch is unset.

pub fn transaction_digest(&self) -> &str

Returns the value of transaction_digest, or the default value if transaction_digest is unset.

pub fn events_digest(&self) -> &str

Returns the value of events_digest, or the default value if events_digest is unset.

pub fn lamport_version(&self) -> u64

Returns the value of lamport_version, or the default value if lamport_version is unset.

pub fn auxiliary_data_digest(&self) -> &str

Returns the value of auxiliary_data_digest, or the default value if auxiliary_data_digest is unset.

impl TransactionEffects

pub const BCS_FIELD: &'static MessageField

pub const DIGEST_FIELD: &'static MessageField

pub const VERSION_FIELD: &'static MessageField

pub const STATUS_FIELD: &'static MessageField

pub const EPOCH_FIELD: &'static MessageField

pub const GAS_USED_FIELD: &'static MessageField

pub const TRANSACTION_DIGEST_FIELD: &'static MessageField

pub const GAS_OBJECT_FIELD: &'static MessageField

pub const EVENTS_DIGEST_FIELD: &'static MessageField

pub const DEPENDENCIES_FIELD: &'static MessageField

pub const LAMPORT_VERSION_FIELD: &'static MessageField

pub const CHANGED_OBJECTS_FIELD: &'static MessageField

pub const UNCHANGED_CONSENSUS_OBJECTS_FIELD: &'static MessageField

pub const AUXILIARY_DATA_DIGEST_FIELD: &'static MessageField

impl TransactionEffects

pub fn path_builder() -> TransactionEffectsFieldPathBuilder

impl TransactionEffects

pub const fn const_default() -> TransactionEffects

pub fn bcs(&self) -> &Bcs

Returns the value of bcs, or the default value if bcs is unset.

pub fn bcs_opt_mut(&mut self) -> Option<&mut Bcs>

If bcs is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn bcs_mut(&mut self) -> &mut Bcs

Returns a mutable reference to bcs. If the field is unset, it is first initialized with the default value.

pub fn bcs_opt(&self) -> Option<&Bcs>

If bcs is set, returns Some with the value; otherwise returns None.

pub fn set_bcs<T>(&mut self, field: T)
where T: Into<Bcs>,

Sets bcs with the provided value.

pub fn with_bcs<T>(self, field: T) -> TransactionEffects
where T: Into<Bcs>,

Sets bcs with the provided value.

pub fn digest_opt_mut(&mut self) -> Option<&mut String>

If digest is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn digest_mut(&mut self) -> &mut String

Returns a mutable reference to digest. If the field is unset, it is first initialized with the default value.

pub fn digest_opt(&self) -> Option<&str>

If digest is set, returns Some with the value; otherwise returns None.

pub fn set_digest<T>(&mut self, field: T)
where T: Into<String>,

Sets digest with the provided value.

pub fn with_digest<T>(self, field: T) -> TransactionEffects
where T: Into<String>,

Sets digest with the provided value.

pub fn version_opt_mut(&mut self) -> Option<&mut i32>

If version is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn version_mut(&mut self) -> &mut i32

Returns a mutable reference to version. If the field is unset, it is first initialized with the default value.

pub fn version_opt(&self) -> Option<i32>

If version is set, returns Some with the value; otherwise returns None.

pub fn set_version<T>(&mut self, field: T)
where T: Into<i32>,

Sets version with the provided value.

pub fn with_version<T>(self, field: T) -> TransactionEffects
where T: Into<i32>,

Sets version with the provided value.

pub fn status(&self) -> &ExecutionStatus

Returns the value of status, or the default value if status is unset.

pub fn status_opt_mut(&mut self) -> Option<&mut ExecutionStatus>

If status is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn status_mut(&mut self) -> &mut ExecutionStatus

Returns a mutable reference to status. If the field is unset, it is first initialized with the default value.

pub fn status_opt(&self) -> Option<&ExecutionStatus>

If status is set, returns Some with the value; otherwise returns None.

pub fn set_status<T>(&mut self, field: T)
where T: Into<ExecutionStatus>,

Sets status with the provided value.

pub fn with_status<T>(self, field: T) -> TransactionEffects
where T: Into<ExecutionStatus>,

Sets status with the provided value.

pub fn epoch_opt_mut(&mut self) -> Option<&mut u64>

If epoch is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn epoch_mut(&mut self) -> &mut u64

Returns a mutable reference to epoch. If the field is unset, it is first initialized with the default value.

pub fn epoch_opt(&self) -> Option<u64>

If epoch is set, returns Some with the value; otherwise returns None.

pub fn set_epoch<T>(&mut self, field: T)
where T: Into<u64>,

Sets epoch with the provided value.

pub fn with_epoch<T>(self, field: T) -> TransactionEffects
where T: Into<u64>,

Sets epoch with the provided value.

pub fn gas_used(&self) -> &GasCostSummary

Returns the value of gas_used, or the default value if gas_used is unset.

pub fn gas_used_opt_mut(&mut self) -> Option<&mut GasCostSummary>

If gas_used is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn gas_used_mut(&mut self) -> &mut GasCostSummary

Returns a mutable reference to gas_used. If the field is unset, it is first initialized with the default value.

pub fn gas_used_opt(&self) -> Option<&GasCostSummary>

If gas_used is set, returns Some with the value; otherwise returns None.

pub fn set_gas_used<T>(&mut self, field: T)
where T: Into<GasCostSummary>,

Sets gas_used with the provided value.

pub fn with_gas_used<T>(self, field: T) -> TransactionEffects
where T: Into<GasCostSummary>,

Sets gas_used with the provided value.

pub fn transaction_digest_opt_mut(&mut self) -> Option<&mut String>

If transaction_digest is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn transaction_digest_mut(&mut self) -> &mut String

Returns a mutable reference to transaction_digest. If the field is unset, it is first initialized with the default value.

pub fn transaction_digest_opt(&self) -> Option<&str>

If transaction_digest is set, returns Some with the value; otherwise returns None.

pub fn set_transaction_digest<T>(&mut self, field: T)
where T: Into<String>,

Sets transaction_digest with the provided value.

pub fn with_transaction_digest<T>(self, field: T) -> TransactionEffects
where T: Into<String>,

Sets transaction_digest with the provided value.

pub fn gas_object(&self) -> &ChangedObject

Returns the value of gas_object, or the default value if gas_object is unset.

pub fn gas_object_opt_mut(&mut self) -> Option<&mut ChangedObject>

If gas_object is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn gas_object_mut(&mut self) -> &mut ChangedObject

Returns a mutable reference to gas_object. If the field is unset, it is first initialized with the default value.

pub fn gas_object_opt(&self) -> Option<&ChangedObject>

If gas_object is set, returns Some with the value; otherwise returns None.

pub fn set_gas_object<T>(&mut self, field: T)
where T: Into<ChangedObject>,

Sets gas_object with the provided value.

pub fn with_gas_object<T>(self, field: T) -> TransactionEffects
where T: Into<ChangedObject>,

Sets gas_object with the provided value.

pub fn events_digest_opt_mut(&mut self) -> Option<&mut String>

If events_digest is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn events_digest_mut(&mut self) -> &mut String

Returns a mutable reference to events_digest. If the field is unset, it is first initialized with the default value.

pub fn events_digest_opt(&self) -> Option<&str>

If events_digest is set, returns Some with the value; otherwise returns None.

pub fn set_events_digest<T>(&mut self, field: T)
where T: Into<String>,

Sets events_digest with the provided value.

pub fn with_events_digest<T>(self, field: T) -> TransactionEffects
where T: Into<String>,

Sets events_digest with the provided value.

pub fn dependencies(&self) -> &[String]

Returns the value of dependencies, or the default value if dependencies is unset.

pub fn dependencies_mut(&mut self) -> &mut Vec<String>

Returns a mutable reference to dependencies. If the field is unset, it is first initialized with the default value.

pub fn set_dependencies(&mut self, field: Vec<String>)

Sets dependencies with the provided value.

pub fn with_dependencies(self, field: Vec<String>) -> TransactionEffects

Sets dependencies with the provided value.

pub fn lamport_version_opt_mut(&mut self) -> Option<&mut u64>

If lamport_version is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn lamport_version_mut(&mut self) -> &mut u64

Returns a mutable reference to lamport_version. If the field is unset, it is first initialized with the default value.

pub fn lamport_version_opt(&self) -> Option<u64>

If lamport_version is set, returns Some with the value; otherwise returns None.

pub fn set_lamport_version<T>(&mut self, field: T)
where T: Into<u64>,

Sets lamport_version with the provided value.

pub fn with_lamport_version<T>(self, field: T) -> TransactionEffects
where T: Into<u64>,

Sets lamport_version with the provided value.

pub fn changed_objects(&self) -> &[ChangedObject]

Returns the value of changed_objects, or the default value if changed_objects is unset.

pub fn changed_objects_mut(&mut self) -> &mut Vec<ChangedObject>

Returns a mutable reference to changed_objects. If the field is unset, it is first initialized with the default value.

pub fn set_changed_objects(&mut self, field: Vec<ChangedObject>)

Sets changed_objects with the provided value.

pub fn with_changed_objects( self, field: Vec<ChangedObject>, ) -> TransactionEffects

Sets changed_objects with the provided value.

pub fn unchanged_consensus_objects(&self) -> &[UnchangedConsensusObject]

Returns the value of unchanged_consensus_objects, or the default value if unchanged_consensus_objects is unset.

pub fn unchanged_consensus_objects_mut( &mut self, ) -> &mut Vec<UnchangedConsensusObject>

Returns a mutable reference to unchanged_consensus_objects. If the field is unset, it is first initialized with the default value.

pub fn set_unchanged_consensus_objects( &mut self, field: Vec<UnchangedConsensusObject>, )

Sets unchanged_consensus_objects with the provided value.

pub fn with_unchanged_consensus_objects( self, field: Vec<UnchangedConsensusObject>, ) -> TransactionEffects

Sets unchanged_consensus_objects with the provided value.

pub fn auxiliary_data_digest_opt_mut(&mut self) -> Option<&mut String>

If auxiliary_data_digest is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn auxiliary_data_digest_mut(&mut self) -> &mut String

Returns a mutable reference to auxiliary_data_digest. If the field is unset, it is first initialized with the default value.

pub fn auxiliary_data_digest_opt(&self) -> Option<&str>

If auxiliary_data_digest is set, returns Some with the value; otherwise returns None.

pub fn set_auxiliary_data_digest<T>(&mut self, field: T)
where T: Into<String>,

Sets auxiliary_data_digest with the provided value.

pub fn with_auxiliary_data_digest<T>(self, field: T) -> TransactionEffects
where T: Into<String>,

Sets auxiliary_data_digest with the provided value.