Module sui_rpc_api::proto::types

Modules

argument

Nested message and enum types in Argument.

changed_object

Nested message and enum types in ChangedObject.

checkpoint_commitment

Nested message and enum types in CheckpointCommitment.

checkpoint_contents

Nested message and enum types in CheckpointContents.

command

Nested message and enum types in Command.

command_argument_error

Nested message and enum types in CommandArgumentError.

consensus_determined_version_assignments

Nested message and enum types in ConsensusDeterminedVersionAssignments.

end_of_epoch_transaction_kind

Nested message and enum types in EndOfEpochTransactionKind.

failure_status

Nested message and enum types in FailureStatus.

input

Nested message and enum types in Input.

move_value

Nested message and enum types in MoveValue.

multisig_member_public_key

Nested message and enum types in MultisigMemberPublicKey.

multisig_member_signature

Nested message and enum types in MultisigMemberSignature.

object_data

Nested message and enum types in ObjectData.

owner

Nested message and enum types in Owner.

package_upgrade_error

Nested message and enum types in PackageUpgradeError.

transaction

Nested message and enum types in Transaction.

transaction_effects

Nested message and enum types in TransactionEffects.

transaction_expiration

Nested message and enum types in TransactionExpiration.

transaction_kind

Nested message and enum types in TransactionKind.

type_argument_error

Nested message and enum types in TypeArgumentError.

type_tag

Nested message and enum types in TypeTag.

unchanged_shared_object

Nested message and enum types in UnchangedSharedObject.

user_signature

Nested message and enum types in UserSignature.

Structs

ActiveJwk

A new JWK.

Address

Unique identifier for an account on the Sui blockchain.

AddressDeniedForCoinError

Address is denied for this coin type.

Argument

An argument to a programmable transaction command.

AuthenticatorStateExpire

Expire old JWKs.

AuthenticatorStateUpdate

Update the set of valid JWKs.

Bcs

Message that represents a type that is serialized and encoded using the BCS format.

Bn254FieldElement

A point on the BN254 elliptic curve.

CancelledTransaction

A transaction that was cancelled.

CancelledTransactions

Set of cancelled transactions.

ChangeEpoch

System transaction used to change the epoch.

ChangedObject

/ Input/output state of an object that was changed during execution.

CheckpointCommitment

A commitment made by a checkpoint.

CheckpointContents

The committed to contents of a checkpoint.

CheckpointSummary

A header for a checkpoint on the Sui blockchain.

CheckpointedTransactionInfo

Transaction information committed to in a checkpoint.

CircomG1

A G1 point.

CircomG2

A G2 point.

Command

A single command in a programmable transaction.

CommandArgumentError

An error with an argument to a command.

CongestedObjectsError

Set of objects that were congested, leading to the transaction’s cancellation.

ConsensusCommitPrologue

Consensus commit prologue system transaction.

ConsensusDeterminedVersionAssignments

Version assignments performed by consensus.

Digest

32-byte output of hashing a Sui structure using the Blake2b256 hash function.

EndOfEpochData

Data, which when included in a CheckpointSummary, signals the end of an Epoch.

EndOfEpochTransaction

Set of operations run at the end of the epoch to close out the current epoch and start the next one.

EndOfEpochTransactionKind

Operation run at the end of an epoch.

Event

An event.

ExecutionStatus

/ The status of an executed transaction.

FailureStatus

An error that can occur during the execution of a transaction.

GasCostSummary

Summary of gas charges.

GasPayment

Payment information for executing a transaction.

GenesisObject

An object part of the initial chain state.

GenesisTransaction

The genesis transaction.

I128

A signed 128-bit integer encoded in little-endian using 16-bytes.

Identifier

A Move identifier.

Input

An input to a user transaction.

Jwk

A JSON web key.

JwkId

Key to uniquely identify a JWK.

MakeMoveVector

Command to build a Move vector out of a set of individual elements.

MergeCoins

Command to merge multiple coins of the same type into a single coin.

ModifiedAtVersion

Indicates that an object was modified at a specific version.

MoveCall

Command to call a Move function.

MoveError

Error that occurred in Move.

MoveField

MoveLocation

/ Location in Move bytecode where an error occurred.s

MoveModule

Module defined by a package.

MovePackage

A Move package.

MoveStruct

A Move struct.

MoveStructValue

MoveValue

MoveVariant

MoveVector

MultisigAggregatedSignature

Aggregated signature from members of a multisig committee.

MultisigCommittee

A multisig committee.

MultisigMember

A member in a multisig committee.

MultisigMemberPublicKey

Set of valid public keys for multisig committee members.

MultisigMemberSignature

A signature from a member of a multisig committee.

NestedResult

An argument type for a nested result.

Object

An object on the Sui blockchain.

ObjectData

Object data, either a package or struct.

ObjectExist

Information about the old version of the object.

ObjectId

Unique identifier for an object on the Sui blockchain.

ObjectReference

Reference to an object.

ObjectReferenceWithOwner

An object reference with owner information.

ObjectWrite

Object write, including all of mutated, created, unwrapped.

Owner

Enum of different types of ownership for an object.

PackageIdDoesNotMatch

Package ID does not match PackageId in upgrade ticket.

PackageUpgradeError

An error with a upgrading a package.

PackageWrite

Package write.

PasskeyAuthenticator

A passkey authenticator.

ProgrammableTransaction

A user transaction.

Publish

Command to publish a new Move package.

RandomnessStateUpdate

Randomness update.

ReadOnlyRoot

Read-only shared object from the input.

RoaringBitmap

A RoaringBitmap. See RoaringFormatSpec for the specification for the serialized format of RoaringBitmaps.

SharedObjectInput

A shared object input.

SimpleSignature

A basic signature.

SizeError

A size error.

SplitCoins

Command to split a single coin object into multiple coins.

StructTag

Type information for a Move struct.

SystemPackage

System package.

Transaction

A transaction.

TransactionEffects

The output or effects of executing a transaction.

TransactionEffectsV1

Version 1 of TransactionEffects.

TransactionEffectsV2

Version 2 of TransactionEffects.

TransactionEvents

Events emitted during the successful execution of a transaction.

TransactionExpiration

A TTL for a transaction.

TransactionKind

Transaction type.

TransferObjects

Command to transfer ownership of a set of objects to an address.

TypeArgumentError

Type argument error.

TypeOrigin

Identifies a struct and the module it was defined in.

TypeTag

Type of a Move value.

U128

An unsigned 128-bit integer encoded in little-endian using 16-bytes.

U256

An unsigned 256-bit integer encoded in little-endian using 32-bytes.

UnchangedSharedObject

A shared object that wasn’t changed during execution.

Upgrade

Command to upgrade an already published package.

UpgradeInfo

/ Upgraded package info for the linkage table.

UserSignature

A signature from a user.

ValidatorAggregatedSignature

/ An aggregated signature from multiple validators.

ValidatorCommittee

The validator set for a particular epoch.

ValidatorCommitteeMember

A member of a validator committee.

VersionAssignment

Object version assignment from consensus.

ZkLoginAuthenticator

A zklogin authenticator.

ZkLoginClaim

A claim of the iss in a zklogin proof.

ZkLoginInputs

A zklogin groth16 proof and the required inputs to perform proof verification.

ZkLoginProof

A zklogin groth16 proof.

ZkLoginPublicIdentifier

Public key equivalent for zklogin authenticators.

Enums

SignatureScheme

Flag use to disambiguate the signature schemes supported by Sui.

Constants

FILE_DESCRIPTOR_SET

Byte encoded FILE_DESCRIPTOR_SET.

Functions

proto_to_timestamp_ms

timestamp_ms_to_proto