Module sui_rpc_api::types

Structs

CheckpointResponse

ExecuteTransactionOptions

Options for the execute transaction endpoint

ExecuteTransactionResponse

Response type for the execute transaction endpoint

FullCheckpointObject

FullCheckpointResponse

FullCheckpointTransaction

GetCheckpointOptions

GetFullCheckpointOptions

GetObjectOptions

GetTransactionOptions

NodeInfo

Basic information about the state of a Node

ObjectResponse

ResolveTransactionQueryParameters

Query parameters for the resolve transaction endpoint

ResolveTransactionResponse

Response type for the execute transaction endpoint

SimulateTransactionQueryParameters

Query parameters for the simulate transaction endpoint

TransactionResponse

TransactionSimulationResponse

Response type for the transaction simulation endpoint

Enums

EffectsFinality

Constants

X_SUI_CHAIN

Chain name of the current chain

X_SUI_CHAIN_ID

Chain ID of the current chain

X_SUI_CHECKPOINT_HEIGHT

Current checkpoint height

X_SUI_CURSOR

Cursor to be used for endpoints that support cursor-based pagination. Pass this to the start field of the endpoint on the next call to get the next page of results.

X_SUI_EPOCH

Current epoch of the chain

X_SUI_LOWEST_AVAILABLE_CHECKPOINT

Lowest available checkpoint for which transaction and checkpoint data can be requested.

X_SUI_LOWEST_AVAILABLE_CHECKPOINT_OBJECTS

Lowest available checkpoint for which object data can be requested.

X_SUI_TIMESTAMP_MS

Current timestamp of the chain - represented as number of milliseconds from the Unix epoch