Module sui_rpc_api::proto::node::v2

Modules

effects_finality

Nested message and enum types in EffectsFinality.

node_service_client

Generated client implementations.

node_service_server

Generated server implementations.

Structs

BalanceChange

The delta, or change, in balance for an address for a particular Coin type.

EffectsFinality

Indicates the finality of the executed transaction.

ExecuteTransactionOptions

DEPRECATED To be removed in the next release

ExecuteTransactionRequest

Request message for NodeService.ExecuteTransaction.

ExecuteTransactionResponse

Response message for NodeService.ExecuteTransaction.

FullCheckpointObject

An object used by or produced from a transaction.

FullCheckpointObjects

Set of objects used by or produced from a transaction. TODO remove next release

FullCheckpointTransaction

A transaction, with all of its inputs and outputs.

GetCheckpointOptions

DEPRECATED To be removed in the next release

GetCheckpointRequest

Request message for NodeService.GetCheckpoint.

GetCheckpointResponse

Response message for NodeService.GetCheckpoint.

GetCommitteeRequest

Request message for NodeService.GetCommittee.

GetCommitteeResponse

Response message for NodeService.GetCommittee.

GetFullCheckpointOptions

DEPRECATED To be removed in the next release

GetFullCheckpointRequest

Request message for NodeService.GetFullCheckpoint.

GetFullCheckpointResponse

Response message for NodeService.GetFullCheckpoint.

GetNodeInfoRequest

Request message for NodeService.GetNodeInfo.

GetNodeInfoResponse

Response message for NodeService.GetNodeInfo.

GetObjectOptions

DEPRECATED To be removed in the next release

GetObjectRequest

Request message for NodeService.GetObject.

GetObjectResponse

Response message for NodeService.GetObject.

GetTransactionOptions

DEPRECATED To be removed in the next release

GetTransactionRequest

Request message for NodeService.GetTransaction.

GetTransactionResponse

Response message for NodeService.GetTransaction.

Constants

FILE_DESCRIPTOR_SET

Byte encoded FILE_DESCRIPTOR_SET.