sui_rpc_api
Struct RpcService

pub struct RpcService { /* private fields */ }

Implementations

impl RpcService

pub fn list_account_objects( &self, request: ListAccountObjectsRequest, ) -> Result<ListAccountObjectsResponse>

impl RpcService

pub fn get_checkpoint( &self, request: GetCheckpointRequest, ) -> Result<GetCheckpointResponse>

pub fn get_full_checkpoint( &self, request: GetFullCheckpointRequest, ) -> Result<GetFullCheckpointResponse>

impl RpcService

pub fn get_coin_info( &self, request: GetCoinInfoRequest, ) -> Result<GetCoinInfoResponse>

impl RpcService

pub fn get_committee( &self, epoch: Option<EpochId>, ) -> Result<ValidatorCommittee>

impl RpcService

pub fn get_gas_info( &self, _request: GetGasInfoRequest, ) -> Result<GetGasInfoResponse>

impl RpcService

pub fn health_check(&self, threshold_seconds: Option<u32>) -> Result<()>

Perform a simple health check on the service.

The threshold, or delta, between the server’s system time and the timestamp in the most recently executed checkpoint for which the server is considered to be healthy.

If not provided, the server will be considered healthy if it can simply fetch the latest checkpoint from its store.

impl RpcService

pub fn get_node_info(&self) -> Result<GetNodeInfoResponse>

impl RpcService

pub fn get_object(&self, _: GetObjectRequest) -> Result<GetObjectResponse>

impl RpcService

pub fn list_dynamic_fields( &self, request: ListDynamicFieldsRequest, ) -> Result<ListDynamicFieldsResponse>

impl RpcService

pub fn get_protocol_config( &self, request: GetProtocolConfigRequest, ) -> Result<GetProtocolConfigResponse>

impl RpcService

pub async fn execute_transaction( &self, request: ExecuteTransactionRequest, ) -> Result<ExecuteTransactionResponse>

pub fn simulate_transaction( &self, parameters: &SimulateTransactionQueryParameters, transaction: Transaction, ) -> Result<TransactionSimulationResponse>

impl RpcService

pub fn resolve_transaction( &self, parameters: ResolveTransactionQueryParameters, unresolved_transaction: Transaction, ) -> Result<ResolveTransactionResponse>

impl RpcService

pub fn get_transaction( &self, _: GetTransactionRequest, ) -> Result<GetTransactionResponse>

impl RpcService

pub fn new( reader: Arc<dyn RpcStateReader>, software_version: &'static str, ) -> Self

pub fn new_without_version(reader: Arc<dyn RpcStateReader>) -> Self

pub fn with_config(&mut self, config: Config)

pub fn with_executor( &mut self, executor: Arc<dyn TransactionExecutor + Send + Sync>, )

pub fn with_subscription_service( &mut self, subscription_service_handle: SubscriptionServiceHandle, )

pub fn with_metrics(&mut self, metrics: RpcMetrics)

pub fn chain_id(&self) -> ChainIdentifier

pub fn software_version(&self) -> &'static str

pub async fn into_router(self) -> Router

pub async fn start_service(self, socket_address: SocketAddr)